+++ packages/message-relayer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @eth-optimism/message-relayer
 
+## 0.5.0
+
+### Minor Changes
+
+- 84a8934c: BaseServiceV2 exposes service name and version as standard synthetic metric
+
+### Patch Changes
+
+- Updated dependencies [d9e39931]
+- Updated dependencies [84a8934c]
+  - @eth-optimism/common-ts@0.3.0
+
 ## 0.4.13
 
 ### Patch Changes